Fix routing.
Fix routing so if you are not login and try to go on another page. It goes to login.
Fix so if you are login, the landing page is /customer.
Fix so if you refresh the page you are not going to login page.

Not fix. If a random URL is writen, a empty page is shown. What is needed to show is didnt find page you where looking for.